desktop/requirements.txt
loguru==0.6.0
paramiko==2.12.0
python-dotenv==0.21.0
requests==2.28.2